Create Silly Game
Alice and Bob play g number of games. Given the value of g which is the number of games.
The integer n for each game is given.
There would be the unique set of n integers from 1 to n.
Alice/ Bob would choose a prime number and remove all its multiples from the set.
Alice starts the game first. The first player who does not have the valid move loses the game.
Print the name of winner.
Note: 1 is not considered as a prime number