Create Prefix Neighbours
Mark has a dictionary, , containing  distinct strings. He defines the benefit value of a string as the sum of the ASCII values of its characters.

Mark calls some string  and some string  prefix neighbors if both of the following conditions are satisfied:
1.  No two strings in the subset are prefix neighbors.
2. The sum of the benefit value of the selected strings is maximal.

String  is a prefix of string .
No other string  exists in the dictionary that is both smaller in length than string  and has string  as a prefix.
For example, in  {"AA", "AB", "ABD", "ABDE"}, the strings AB and ABD are prefix neighbors because AB is a prefix of ABD and no other string of length  in  shares that prefix. The strings ABD and ABDE are also prefix neighbors, but AB and ABDE are not prefix neighbors because string  exists (i.e., ABD).

Mark wants to select a subset of strings from  such that the following conditions are both satisfied: